Initial commit of the Virtual Pointer
This initial commit implements the basic structure of the virtual
pointer interface for buffers, that enables allocation of deallocation
of pointer-like structures that are not dereferenciable on the host
but can be used to retrieve buffers and offsets later on.
PointerMapper is a normal object, unlike the singleton implemented for
the legacy pointer interface.
See README.adoc for details.